StarRating improvements

+++ RandomRestaurantQuizz.Core/Quizzz/Events/ScoreChangedEvent.cs
@@ -2,13 +2,13 @@
 
 public class ScoreChangedEvent
 {
-    public ScoreChangedEvent(int totalScore, double roundScore, double locationRating)
+    public ScoreChangedEvent(double totalScore, double roundScore, double locationRating)
     {
         TotalScore = totalScore;
         RoundScore = roundScore;
         ScoreDiff = $" +{roundScore} ({locationRating:n1})";
     }
-    public int TotalScore { get; }
+    public double TotalScore { get; }
     public double RoundScore { get; }
     public string ScoreDiff { get; }
 }
